# Remove a Barrett reduction during keygen and decryption