--- a/ref/poly.c
+++ b/ref/poly.c
@@ -243,7 +243,6 @@ void poly_getnoise(poly *r, const uint8_t seed[KYBER_SYMBYTES], uint8_t nonce)
 void poly_ntt(poly *r)
 {
   ntt(r->coeffs);
-  poly_reduce(r);
 }
 
 /*************************************************