For keygen, coefficients of e are bounded in absolute value by 7q (NTT). Those of
pkpv are bounded in absolute value by q (to_mont).  Thus Barrett
reducing e before adding it to pkpv is unnecessary.

For decryption skpv is bounded in absolute value by 2^12.  bp is bounded in absolute
value by 7q.  Multiplied the coefficients are bounded by 2^14.9q,
which is below the 2^15q required for the Montgomery reduction.

This only changes the ref code.  I'll leave the avx2 changes to you.