@kaepora 2020-08-28
This seems like a reasonable change but keep in mind that the performance improvements seem minute. In https://github.com/symbolicsoft/kyber-k2so, we obtain the following benchmark results:

Before:
```
goos: linux
goarch: amd64
pkg: github.com/symbolicsoft/kyber-k2so
BenchmarkKemKeypair512-8           24496             48893 ns/op
BenchmarkKemKeypair768-8           14337             84809 ns/op
BenchmarkKemKeypair1024-8           9048            128939 ns/op
BenchmarkKemEncrypt512-8           18960             62227 ns/op
BenchmarkKemEncrypt768-8           10000            102490 ns/op
BenchmarkKemEncrypt1024-8           7736            149436 ns/op
BenchmarkKemDecrypt512-8           16453             71740 ns/op
BenchmarkKemDecrypt768-8           10000            115441 ns/op
BenchmarkKemDecrypt1024-8           6794            172428 ns/op
PASS
ok      github.com/symbolicsoft/kyber-k2so      14.464s
```

After:
```
goos: linux
goarch: amd64
pkg: github.com/symbolicsoft/kyber-k2so
BenchmarkKemKeypair512-8           24832             48027 ns/op
BenchmarkKemKeypair768-8           14223             83849 ns/op
BenchmarkKemKeypair1024-8           8811            128995 ns/op
BenchmarkKemEncrypt512-8           19198             61744 ns/op
BenchmarkKemEncrypt768-8           10000            100608 ns/op
BenchmarkKemEncrypt1024-8           6814            150601 ns/op
BenchmarkKemDecrypt512-8           16560             71019 ns/op
BenchmarkKemDecrypt768-8            9692            113459 ns/op
BenchmarkKemDecrypt1024-8           6862            166649 ns/op
PASS
ok      github.com/symbolicsoft/kyber-k2so      14.123s
```