@gregorseiler 2020-12-07
Thanks for pointing this out. For the ref code, which puts emphasis on simplicity, I think it's better not to have functions that are delicate to use because of large outputs and keep everything nicely bounded. For the round 3 code I have actually revisited the range analysis and the avx2 code now does something similar than what you have suggested.